Adds reset seed usage
